Feat: 로그인/로그아웃하면 홈페이지 리렌더링 되도록
