+++ src/pages/homepage/homePage.tsx
@@ -42,7 +42,7 @@ const HomePage: React.FC = () => {
     };
 
     fetchHomeData();
-  }, []);
+  }, [isAuthenticated]);
 
   if (!homeData) return <div>Loading...</div>;
